wip painter pane cleanup
