+++ src/components/controls/PainterPane/index.ts
@@ -1,0 +1,1 @@
+export { default } from './PainterPane';
