+++ src/data/campaigns/index.ts
@@ -1,12 +1,24 @@
-import { Campaign } from '../../types/Campaign';
 import mortal from './mortal';
 import vortex from './vortex';
-
-type Campaigns = Readonly<{ [K in keyof typeof campaigns]: Campaign }>;
 
 const campaigns = Object.freeze({
   mortal,
   vortex,
 });
 
-export default campaigns as Campaigns;
+export interface Campaign {
+  readonly key: string;
+  readonly name: string;
+  readonly map: {
+    readonly image: string;
+    readonly imageText: string;
+    readonly width: number;
+    readonly height: number;
+  };
+  readonly regions: any;
+}
+
+export type CampaignKey = keyof typeof campaigns;
+export type CampaignLookup = Readonly<{ [key in CampaignKey]: Campaign }>;
+
+export default campaigns as CampaignLookup;
