--- src/types/Campaign.ts
@@ -1,11 +1,0 @@
-export interface Campaign {
-  readonly key: string;
-  readonly name: string;
-  readonly map: {
-    readonly image: string;
-    readonly imageText: string;
-    readonly width: number;
-    readonly height: number;
-  };
-  readonly regions: any;
-}
